Client

This page lists docstrings relevant to the client API.

Adding different types of nodes:

Open62541.JUA_Client_addNode — Function

JUA_Client_addNode(client::JUA_Client, requestedNewNodeId::JUA_NodeId,
        parentNodeId::JUA_NodeId, referenceTypeId::JUA_NodeId, browseName::JUA_QualifiedName, 
        attributes::Union{JUA_VariableAttributes, JUA_ObjectAttributes},
        outNewNodeId::JUA_NodeId, typeDefinition::JUA_NodeId)::UA_StatusCode

uses the client API to add a Variable or Object node to the server to which the client is connected to.

See JUA_VariableAttributes, JUA_VariableTypeAttributes, and JUA_ObjectAttributes on how to define valid attributes.

JUA_Client_addNode(client::JUA_Client, requestedNewNodeId::JUA_NodeId,
        parentNodeId, referenceTypeId::JUA_NodeId, browseName::JUA_QualifiedName,
        attributes::Union{JUA_VariableTypeAttributes, JUA_ObjectTypeAttributes, 
        JUA_ReferenceTypeAttributes, JUA_DataTypeAttributes, JUA_ViewAttributes},
        outNewNodeId::JUA_NodeId)::UA_StatusCode

uses the client API to add a ObjectType, ReferenceType, DataType or View node to the server to which the client is connected to.

See `JUA_VariableTypeAttributes, JUA_ObjectTypeAttributes, JUA_ReferenceTypeAttributes, JUA_DataTypeAttributes, and JUA_ViewAttributes on how to define valid attributes.

Open62541.UA_Client_addVariableNode — Function

UA_Client_addVariableNode(::Ptr{UA_Client}, requestednewnodeid::Ptr{UA_NodeId}, 
        parentnodeid::Ptr{UA_NodeId}, referenceTypeId::Ptr{UA_NodeId}, 
        browseName::Ptr{UA_QualifiedName}, typedefinition::Ptr{UA_NodeId},
        attr::Ptr{UA_VariableAttributes}, outNewNodeId::Ptr{UA_NodeId})::UA_StatusCode

uses the client API to add a variable node to the client.

See UA_VariableAttributes_generate on how to define valid attributes.

Open62541.UA_Client_addObjectNode — Function

UA_Client_addObjectNode(::Ptr{UA_Client}, requestednewnodeid::Ptr{UA_NodeId}, 
        parentnodeid::Ptr{UA_NodeId}, referenceTypeId::Ptr{UA_NodeId}, 
        browseName::Ptr{UA_QualifiedName}, typedefinition::Ptr{UA_NodeId},
        attr::Ptr{UA_ObjectAttributes}, outNewNodeId::Ptr{UA_NodeId})::UA_StatusCode

uses the client API to add a object node to the client.

See UA_ObjectAttributes_generate on how to define valid attributes.

Open62541.UA_Client_addVariableTypeNode — Function

UA_Client_addVariableTypeNode(::Ptr{UA_Client}, requestednewnodeid::Ptr{UA_NodeId}, 
        parentnodeid::Ptr{UA_NodeId}, referenceTypeId::Ptr{UA_NodeId}, 
        browseName::Ptr{UA_QualifiedName}, attr::Ptr{UA_VariableTypeAttributes}, 
        outNewNodeId::Ptr{UA_NodeId})::UA_StatusCode

uses the client API to add a variabletype node to the client.

See UA_VariableTypeAttributes_generate on how to define valid attributes.

Open62541.UA_Client_addObjectTypeNode — Function

UA_Client_addObjectTypeNode(::Ptr{UA_Client}, requestednewnodeid::Ptr{UA_NodeId}, 
        parentnodeid::Ptr{UA_NodeId}, referenceTypeId::Ptr{UA_NodeId}, 
        browseName::Ptr{UA_QualifiedName}, attr::Ptr{UA_ObjectTypeAttributes}, 
        outNewNodeId::Ptr{UA_NodeId})::UA_StatusCode

uses the client API to add a objecttype node to the client.

See UA_ObjectTypeAttributes_generate on how to define valid attributes.

Open62541.UA_Client_addViewNode — Function

UA_Client_addViewNode(::Ptr{UA_Client}, requestednewnodeid::Ptr{UA_NodeId}, 
        parentnodeid::Ptr{UA_NodeId}, referenceTypeId::Ptr{UA_NodeId}, 
        browseName::Ptr{UA_QualifiedName}, attr::Ptr{UA_ViewAttributes}, 
        outNewNodeId::Ptr{UA_NodeId})::UA_StatusCode

uses the client API to add a view node to the client.

See UA_ViewAttributes_generate on how to define valid attributes.

Open62541.UA_Client_addReferenceTypeNode — Function

UA_Client_addReferenceTypeNode(::Ptr{UA_Client}, requestednewnodeid::Ptr{UA_NodeId}, 
        parentnodeid::Ptr{UA_NodeId}, referenceTypeId::Ptr{UA_NodeId}, 
        browseName::Ptr{UA_QualifiedName}, attr::Ptr{UA_ReferenceTypeAttributes}, 
        outNewNodeId::Ptr{UA_NodeId})::UA_StatusCode

uses the client API to add a referencetype node to the client.

See UA_ReferenceTypeAttributes_generate on how to define valid attributes.

Open62541.UA_Client_addDataTypeNode — Function

UA_Client_addDataTypeNode(::Ptr{UA_Client}, requestednewnodeid::Ptr{UA_NodeId}, 
        parentnodeid::Ptr{UA_NodeId}, referenceTypeId::Ptr{UA_NodeId}, 
        browseName::Ptr{UA_QualifiedName}, attr::Ptr{UA_DataTypeAttributes}, 
        outNewNodeId::Ptr{UA_NodeId})::UA_StatusCode

uses the client API to add a datatype node to the client.

See UA_DataTypeAttributes_generate on how to define valid attributes.

Reading from nodes:

Open62541.JUA_Client_readValueAttribute — Function

value = JUA_Client_readValueAttribute(client::JUA_Client, nodeId::JUA_NodeId, type = Any)

uses the client API to read the value of nodeId from the server that the client is connected to.

The output value is automatically converted to a Julia type (such as Float64, String, Vector{String}, etc.) if possible. Otherwise, open62541 composite types are returned.

Note: Since it is unknown what type of value is stored within nodeId before reading it, this function is inherently type unstable.

Type stability is improved if the optional argument type is provided, for example, if you know that you have stored a Matrix{Float64} in nodeId, then you should specify this. If the wrong type is specified, the function will throw a TypeError.

Open62541.UA_Client_readAccessLevelAttribute — Function

UA_Client_readAccessLevelAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_Byte})

Uses the UA Client API to read the value of attribute AccessLevel from the NodeId nodeId accessed through the client client.

Open62541.UA_Client_readBrowseNameAttribute — Function

UA_Client_readBrowseNameAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_QualifiedName})

Uses the UA Client API to read the value of attribute BrowseName from the NodeId nodeId accessed through the client client.

Open62541.UA_Client_readContainsNoLoopsAttribute — Function

UA_Client_readContainsNoLoopsAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_Boolean})

Uses the UA Client API to read the value of attribute ContainsNoLoops from the NodeId nodeId accessed through the client client.

Open62541.UA_Client_readDataTypeAttribute — Function

UA_Client_readDataTypeAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_NodeId})

Uses the UA Client API to read the value of attribute DataType from the NodeId nodeId accessed through the client client.

Open62541.UA_Client_readDescriptionAttribute — Function

UA_Client_readDescriptionAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_LocalizedText})

Uses the UA Client API to read the value of attribute Description from the NodeId nodeId accessed through the client client.

Open62541.UA_Client_readDisplayNameAttribute — Function

UA_Client_readDisplayNameAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_LocalizedText})

Uses the UA Client API to read the value of attribute DisplayName from the NodeId nodeId accessed through the client client.

Open62541.UA_Client_readEventNotifierAttribute — Function

UA_Client_readEventNotifierAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_Byte})

Uses the UA Client API to read the value of attribute EventNotifier from the NodeId nodeId accessed through the client client.

Open62541.UA_Client_readExecutableAttribute — Function

UA_Client_readExecutableAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_Boolean})

Uses the UA Client API to read the value of attribute Executable from the NodeId nodeId accessed through the client client.

Open62541.UA_Client_readHistorizingAttribute — Function

UA_Client_readHistorizingAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_Boolean})

Uses the UA Client API to read the value of attribute Historizing from the NodeId nodeId accessed through the client client.

Open62541.UA_Client_readInverseNameAttribute — Function

UA_Client_readInverseNameAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_LocalizedText})

Uses the UA Client API to read the value of attribute InverseName from the NodeId nodeId accessed through the client client.

Open62541.UA_Client_readIsAbstractAttribute — Function

UA_Client_readIsAbstractAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_Boolean})

Uses the UA Client API to read the value of attribute IsAbstract from the NodeId nodeId accessed through the client client.

Open62541.UA_Client_readMinimumSamplingIntervalAttribute — Function

UA_Client_readMinimumSamplingIntervalAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_Double})

Uses the UA Client API to read the value of attribute MinimumSamplingInterval from the NodeId nodeId accessed through the client client.

Open62541.UA_Client_readNodeClassAttribute — Function

UA_Client_readNodeClassAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_NodeClass})

Uses the UA Client API to read the value of attribute NodeClass from the NodeId nodeId accessed through the client client.

Open62541.UA_Client_readNodeIdAttribute — Function

UA_Client_readNodeIdAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_NodeId})

Uses the UA Client API to read the value of attribute NodeId from the NodeId nodeId accessed through the client client.

Open62541.UA_Client_readSymmetricAttribute — Function

UA_Client_readSymmetricAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_Boolean})

Uses the UA Client API to read the value of attribute Symmetric from the NodeId nodeId accessed through the client client.

Open62541.UA_Client_readUserAccessLevelAttribute — Function

UA_Client_readUserAccessLevelAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_Byte})

Uses the UA Client API to read the value of attribute UserAccessLevel from the NodeId nodeId accessed through the client client.

Open62541.UA_Client_readUserExecutableAttribute — Function

UA_Client_readUserExecutableAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_Boolean})

Uses the UA Client API to read the value of attribute UserExecutable from the NodeId nodeId accessed through the client client.

Open62541.UA_Client_readUserWriteMaskAttribute — Function

UA_Client_readUserWriteMaskAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_UInt32})

Uses the UA Client API to read the value of attribute UserWriteMask from the NodeId nodeId accessed through the client client.

Open62541.UA_Client_readValueAttribute — Function

UA_Client_readValueAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_Variant})

Uses the UA Client API to read the value of attribute Value from the NodeId nodeId accessed through the client client.

Open62541.UA_Client_readValueRankAttribute — Function

UA_Client_readValueRankAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_Int32})

Uses the UA Client API to read the value of attribute ValueRank from the NodeId nodeId accessed through the client client.

Open62541.UA_Client_readWriteMaskAttribute — Function

UA_Client_readWriteMaskAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_UInt32})

Uses the UA Client API to read the value of attribute WriteMask from the NodeId nodeId accessed through the client client.

Writing to nodes:

Open62541.JUA_Client_writeValueAttribute — Function

JUA_Client_writeValueAttribute(server::JUA_Client, nodeId::JUA_NodeId, newvalue)::UA_StatusCode

uses the client API to write the value newvalue to nodeId to the server that the client is connected to.

new_value must either be a JUA_Variant or a Julia value/array compatible with any of its constructors.

See also JUA_Variant

Open62541.UA_Client_writeAccessLevelAttribute — Function

UA_Client_writeAccessLevelAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_Byte})

Uses the UA Client API to write the value new_val to the attribute AccessLevel of the NodeId nodeId accessed through the client client.

Open62541.UA_Client_writeBrowseNameAttribute — Function

UA_Client_writeBrowseNameAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_QualifiedName})

Uses the UA Client API to write the value new_val to the attribute BrowseName of the NodeId nodeId accessed through the client client.

Open62541.UA_Client_writeContainsNoLoopsAttribute — Function

UA_Client_writeContainsNoLoopsAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_Boolean})

Uses the UA Client API to write the value new_val to the attribute ContainsNoLoops of the NodeId nodeId accessed through the client client.

Open62541.UA_Client_writeDataTypeAttribute — Function

UA_Client_writeDataTypeAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_NodeId})

Uses the UA Client API to write the value new_val to the attribute DataType of the NodeId nodeId accessed through the client client.

Open62541.UA_Client_writeDescriptionAttribute — Function

UA_Client_writeDescriptionAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_LocalizedText})

Uses the UA Client API to write the value new_val to the attribute Description of the NodeId nodeId accessed through the client client.

Open62541.UA_Client_writeDisplayNameAttribute — Function

UA_Client_writeDisplayNameAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_LocalizedText})

Uses the UA Client API to write the value new_val to the attribute DisplayName of the NodeId nodeId accessed through the client client.

Open62541.UA_Client_writeEventNotifierAttribute — Function

UA_Client_writeEventNotifierAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_Byte})

Uses the UA Client API to write the value new_val to the attribute EventNotifier of the NodeId nodeId accessed through the client client.

Open62541.UA_Client_writeExecutableAttribute — Function

UA_Client_writeExecutableAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_Boolean})

Uses the UA Client API to write the value new_val to the attribute Executable of the NodeId nodeId accessed through the client client.

Open62541.UA_Client_writeHistorizingAttribute — Function

UA_Client_writeHistorizingAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_Boolean})

Uses the UA Client API to write the value new_val to the attribute Historizing of the NodeId nodeId accessed through the client client.

Open62541.UA_Client_writeInverseNameAttribute — Function

UA_Client_writeInverseNameAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_LocalizedText})

Uses the UA Client API to write the value new_val to the attribute InverseName of the NodeId nodeId accessed through the client client.

Open62541.UA_Client_writeIsAbstractAttribute — Function

UA_Client_writeIsAbstractAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_Boolean})

Uses the UA Client API to write the value new_val to the attribute IsAbstract of the NodeId nodeId accessed through the client client.

Open62541.UA_Client_writeMinimumSamplingIntervalAttribute — Function

UA_Client_writeMinimumSamplingIntervalAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_Double})

Uses the UA Client API to write the value new_val to the attribute MinimumSamplingInterval of the NodeId nodeId accessed through the client client.

Open62541.UA_Client_writeNodeClassAttribute — Function

UA_Client_writeNodeClassAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_NodeClass})

Uses the UA Client API to write the value new_val to the attribute NodeClass of the NodeId nodeId accessed through the client client.

Open62541.UA_Client_writeNodeIdAttribute — Function

UA_Client_writeNodeIdAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_NodeId})

Uses the UA Client API to write the value new_val to the attribute NodeId of the NodeId nodeId accessed through the client client.

Open62541.UA_Client_writeSymmetricAttribute — Function

UA_Client_writeSymmetricAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_Boolean})

Uses the UA Client API to write the value new_val to the attribute Symmetric of the NodeId nodeId accessed through the client client.

Open62541.UA_Client_writeUserAccessLevelAttribute — Function

UA_Client_writeUserAccessLevelAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_Byte})

Uses the UA Client API to write the value new_val to the attribute UserAccessLevel of the NodeId nodeId accessed through the client client.

Open62541.UA_Client_writeUserExecutableAttribute — Function

UA_Client_writeUserExecutableAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_Boolean})

Uses the UA Client API to write the value new_val to the attribute UserExecutable of the NodeId nodeId accessed through the client client.

Open62541.UA_Client_writeUserWriteMaskAttribute — Function

UA_Client_writeUserWriteMaskAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_UInt32})

Uses the UA Client API to write the value new_val to the attribute UserWriteMask of the NodeId nodeId accessed through the client client.

Open62541.UA_Client_writeValueAttribute — Function

UA_Client_writeValueAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_Variant})

Uses the UA Client API to write the value new_val to the attribute Value of the NodeId nodeId accessed through the client client.

Open62541.UA_Client_writeValueRankAttribute — Function

UA_Client_writeValueRankAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_Int32})

Uses the UA Client API to write the value new_val to the attribute ValueRank of the NodeId nodeId accessed through the client client.

Open62541.UA_Client_writeWriteMaskAttribute — Function

UA_Client_writeWriteMaskAttribute(client::Ptr{UA_Client}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_UInt32})

Uses the UA Client API to write the value new_val to the attribute WriteMask of the NodeId nodeId accessed through the client client.

Asynchronous Client API