Server

This page lists docstrings relevant to the server API.

Adding different types of nodes:

Open62541.JUA_Server_addNode — Function

JUA_Server_addNode(server::JUA_Server, requestedNewNodeId::JUA_NodeId,
        parentNodeId::JUA_NodeId, referenceTypeId::JUA_NodeId, browseName::JUA_QualifiedName, 
        attributes::Union{JUA_VariableAttributes, JUA_VariableTypeAttributes, JUA_ObjectAttributes},
        outNewNodeId::JUA_NodeId, nodeContext::JUA_NodeId, typeDefinition::JUA_NodeId)::UA_StatusCode

uses the server API to add a Variable, VariableType, or Object node to the server.

See JUA_VariableAttributes, JUA_VariableTypeAttributes, and JUA_ObjectAttributes on how to define valid attributes.

JUA_Server_addNode(server::JUA_Server, requestedNewNodeId::JUA_NodeId,
        parentNodeId, referenceTypeId::JUA_NodeId, browseName::JUA_QualifiedName,
        attributes::Union{JUA_ObjectTypeAttributes, JUA_ReferenceTypeAttributes, JUA_DataTypeAttributes, JUA_ViewAttributes},
        outNewNodeId::JUA_NodeId, nodeContext::JUA_NodeId)::UA_StatusCode

uses the server API to add a ObjectType, ReferenceType, DataType, or View node to the server.

See JUA_ObjectTypeAttributes, JUA_ReferenceTypeAttributes, JUA_DataTypeAttributes, and JUA_ViewAttributes on how to define valid attributes.

TODO: Need to add docstring for method node addition once I have thought about the interface.

Open62541.UA_Server_addVariableNode — Function

UA_Server_addVariableNode(server::Ptr{UA_Server}, requestednewnodeid::Ptr{UA_NodeId}, 
        parentnodeid::Ptr{UA_NodeId}, referenceTypeId::Ptr{UA_NodeId}, 
        browseName::Ptr{UA_QualifiedName}, typedefinition::Ptr{UA_NodeId},
        attr::Ptr{UA_VariableAttributes}, nodeContext::Ptr{UA_NodeId}, 
        outNewNodeId::Ptr{UA_NodeId})::UA_StatusCode

uses the server API to add a variable node to the server.

See UA_VariableAttributes_generate on how to define valid attributes.

Open62541.UA_Server_addObjectNode — Function

UA_Server_addObjectNode(server::Ptr{UA_Server}, requestednewnodeid::Ptr{UA_NodeId}, 
        parentnodeid::Ptr{UA_NodeId}, referenceTypeId::Ptr{UA_NodeId}, 
        browseName::Ptr{UA_QualifiedName}, typedefinition::Ptr{UA_NodeId},
        attr::Ptr{UA_ObjectAttributes}, nodeContext::Ptr{UA_NodeId}, 
        outNewNodeId::Ptr{UA_NodeId})::UA_StatusCode

uses the server API to add a object node to the server.

See UA_ObjectAttributes_generate on how to define valid attributes.

Open62541.UA_Server_addVariableTypeNode — Function

UA_Server_addVariableTypeNode(server::Ptr{UA_Server}, requestednewnodeid::Ptr{UA_NodeId}, 
        parentnodeid::Ptr{UA_NodeId}, referenceTypeId::Ptr{UA_NodeId}, 
        browseName::Ptr{UA_QualifiedName}, typedefinition::Ptr{UA_NodeId},
        attr::Ptr{UA_VariableTypeAttributes}, nodeContext::Ptr{UA_NodeId}, 
        outNewNodeId::Ptr{UA_NodeId})::UA_StatusCode

uses the server API to add a variabletype node to the server.

See UA_VariableTypeAttributes_generate on how to define valid attributes.

Open62541.UA_Server_addObjectTypeNode — Function

UA_Server_addObjectTypeNode(server::Ptr{UA_Server}, requestednewnodeid::Ptr{UA_NodeId}, 
        parentnodeid::Ptr{UA_NodeId}, referenceTypeId::Ptr{UA_NodeId}, 
        browseName::Ptr{UA_QualifiedName}, typedefinition::Ptr{UA_NodeId},
        attr::Ptr{UA_ObjectTypeAttributes}, nodeContext::Ptr{UA_NodeId}, 
        outNewNodeId::Ptr{UA_NodeId})::UA_StatusCode

uses the server API to add a objecttype node to the server.

See UA_ObjectTypeAttributes_generate on how to define valid attributes.

Open62541.UA_Server_addViewNode — Function

UA_Server_addViewNode(server::Ptr{UA_Server}, requestednewnodeid::Ptr{UA_NodeId}, 
        parentnodeid::Ptr{UA_NodeId}, referenceTypeId::Ptr{UA_NodeId}, 
        browseName::Ptr{UA_QualifiedName}, typedefinition::Ptr{UA_NodeId},
        attr::Ptr{UA_ViewAttributes}, nodeContext::Ptr{UA_NodeId}, 
        outNewNodeId::Ptr{UA_NodeId})::UA_StatusCode

uses the server API to add a view node to the server.

See UA_ViewAttributes_generate on how to define valid attributes.

Open62541.UA_Server_addReferenceTypeNode — Function

UA_Server_addReferenceTypeNode(server::Ptr{UA_Server}, requestednewnodeid::Ptr{UA_NodeId}, 
        parentnodeid::Ptr{UA_NodeId}, referenceTypeId::Ptr{UA_NodeId}, 
        browseName::Ptr{UA_QualifiedName}, typedefinition::Ptr{UA_NodeId},
        attr::Ptr{UA_ReferenceTypeAttributes}, nodeContext::Ptr{UA_NodeId}, 
        outNewNodeId::Ptr{UA_NodeId})::UA_StatusCode

uses the server API to add a referencetype node to the server.

See UA_ReferenceTypeAttributes_generate on how to define valid attributes.

Open62541.UA_Server_addDataTypeNode — Function

UA_Server_addDataTypeNode(server::Ptr{UA_Server}, requestednewnodeid::Ptr{UA_NodeId}, 
        parentnodeid::Ptr{UA_NodeId}, referenceTypeId::Ptr{UA_NodeId}, 
        browseName::Ptr{UA_QualifiedName}, typedefinition::Ptr{UA_NodeId},
        attr::Ptr{UA_DataTypeAttributes}, nodeContext::Ptr{UA_NodeId}, 
        outNewNodeId::Ptr{UA_NodeId})::UA_StatusCode

uses the server API to add a datatype node to the server.

See UA_DataTypeAttributes_generate on how to define valid attributes.

Open62541.UA_Server_addMethodNode — Function

UA_Server_addMethodNode(server::Ptr{UA_Server}, requestednewnodeid::Ptr{UA_NodeId}, 
        parentnodeid::Ptr{UA_NodeId}, referenceTypeId::Ptr{UA_NodeId}, 
        browseName::Ptr{UA_QualifiedName}, attr::Ptr{UA_MethodAttributes}, 
        method::UA_MethodCallback, inputArgumentsSize::Csize_t, inputArguments::Union{UA_Argument, AbstractArray{UA_Argument}}, 
        outputArgumentsSize::Csize_t, outputArguments::Union{UA_Argument, AbstractArray{UA_Argument}}, 
        nodeContext::Ptr{UA_NodeId}, outNewNodeId::Ptr{UA_NodeId})::UA_StatusCode

uses the server API to add a method node with the callback method to the server.

See also:

UA_MethodAttributes_generate to define valid attributes.

UA_MethodAttributes_generate to generate a valid callback.

Reading from nodes:

Open62541.UA_Server_readAccessLevel — Function

UA_Server_readAccessLevel(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_Byte})

Uses the Server API to read the value of the attribute AccessLevel from the NodeId nodeId located on server server. The result is saved into out.

Note that memory for out must be allocated by C before using this function. This can be accomplished with out = UA_Byte_new(). The resulting object must be cleaned up via UA_Byte_delete(out::Ptr{UA_Byte}) after its use.

Open62541.UA_Server_readArrayDimensions — Function

UA_Server_readArrayDimensions(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_Variant})

Uses the Server API to read the value of the attribute ArrayDimensions from the NodeId nodeId located on server server. The result is saved into out.

Note that memory for out must be allocated by C before using this function. This can be accomplished with out = UA_Variant_new(). The resulting object must be cleaned up via UA_Variant_delete(out::Ptr{UA_Variant}) after its use.

Open62541.UA_Server_readBrowseName — Function

UA_Server_readBrowseName(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_QualifiedName})

Uses the Server API to read the value of the attribute BrowseName from the NodeId nodeId located on server server. The result is saved into out.

Note that memory for out must be allocated by C before using this function. This can be accomplished with out = UA_QualifiedName_new(). The resulting object must be cleaned up via UA_QualifiedName_delete(out::Ptr{UA_QualifiedName}) after its use.

Open62541.UA_Server_readContainsNoLoops — Function

UA_Server_readContainsNoLoops(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_Boolean})

Uses the Server API to read the value of the attribute ContainsNoLoops from the NodeId nodeId located on server server. The result is saved into out.

Note that memory for out must be allocated by C before using this function. This can be accomplished with out = UA_Boolean_new(). The resulting object must be cleaned up via UA_Boolean_delete(out::Ptr{UA_Boolean}) after its use.

Open62541.UA_Server_readDataType — Function

UA_Server_readDataType(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_NodeId})

Uses the Server API to read the value of the attribute DataType from the NodeId nodeId located on server server. The result is saved into out.

Note that memory for out must be allocated by C before using this function. This can be accomplished with out = UA_NodeId_new(). The resulting object must be cleaned up via UA_NodeId_delete(out::Ptr{UA_NodeId}) after its use.

Open62541.UA_Server_readDescription — Function

UA_Server_readDescription(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_LocalizedText})

Uses the Server API to read the value of the attribute Description from the NodeId nodeId located on server server. The result is saved into out.

Note that memory for out must be allocated by C before using this function. This can be accomplished with out = UA_LocalizedText_new(). The resulting object must be cleaned up via UA_LocalizedText_delete(out::Ptr{UA_LocalizedText}) after its use.

Open62541.UA_Server_readDisplayName — Function

UA_Server_readDisplayName(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_LocalizedText})

Uses the Server API to read the value of the attribute DisplayName from the NodeId nodeId located on server server. The result is saved into out.

Note that memory for out must be allocated by C before using this function. This can be accomplished with out = UA_LocalizedText_new(). The resulting object must be cleaned up via UA_LocalizedText_delete(out::Ptr{UA_LocalizedText}) after its use.

Open62541.UA_Server_readEventNotifier — Function

UA_Server_readEventNotifier(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_Byte})

Uses the Server API to read the value of the attribute EventNotifier from the NodeId nodeId located on server server. The result is saved into out.

Note that memory for out must be allocated by C before using this function. This can be accomplished with out = UA_Byte_new(). The resulting object must be cleaned up via UA_Byte_delete(out::Ptr{UA_Byte}) after its use.

Open62541.UA_Server_readExecutable — Function

UA_Server_readExecutable(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_Boolean})

Uses the Server API to read the value of the attribute Executable from the NodeId nodeId located on server server. The result is saved into out.

Note that memory for out must be allocated by C before using this function. This can be accomplished with out = UA_Boolean_new(). The resulting object must be cleaned up via UA_Boolean_delete(out::Ptr{UA_Boolean}) after its use.

Open62541.UA_Server_readHistorizing — Function

UA_Server_readHistorizing(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_Boolean})

Uses the Server API to read the value of the attribute Historizing from the NodeId nodeId located on server server. The result is saved into out.

Note that memory for out must be allocated by C before using this function. This can be accomplished with out = UA_Boolean_new(). The resulting object must be cleaned up via UA_Boolean_delete(out::Ptr{UA_Boolean}) after its use.

Open62541.UA_Server_readInverseName — Function

UA_Server_readInverseName(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_LocalizedText})

Uses the Server API to read the value of the attribute InverseName from the NodeId nodeId located on server server. The result is saved into out.

Note that memory for out must be allocated by C before using this function. This can be accomplished with out = UA_LocalizedText_new(). The resulting object must be cleaned up via UA_LocalizedText_delete(out::Ptr{UA_LocalizedText}) after its use.

Open62541.UA_Server_readIsAbstract — Function

UA_Server_readIsAbstract(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_Boolean})

Uses the Server API to read the value of the attribute IsAbstract from the NodeId nodeId located on server server. The result is saved into out.

Note that memory for out must be allocated by C before using this function. This can be accomplished with out = UA_Boolean_new(). The resulting object must be cleaned up via UA_Boolean_delete(out::Ptr{UA_Boolean}) after its use.

Open62541.UA_Server_readMinimumSamplingInterval — Function

UA_Server_readMinimumSamplingInterval(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_Double})

Uses the Server API to read the value of the attribute MinimumSamplingInterval from the NodeId nodeId located on server server. The result is saved into out.

Note that memory for out must be allocated by C before using this function. This can be accomplished with out = UA_Double_new(). The resulting object must be cleaned up via UA_Double_delete(out::Ptr{UA_Double}) after its use.

Open62541.UA_Server_readNodeClass — Function

UA_Server_readNodeClass(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_NodeClass})

Uses the Server API to read the value of the attribute NodeClass from the NodeId nodeId located on server server. The result is saved into out.

Note that memory for out must be allocated by C before using this function. This can be accomplished with out = UA_NodeClass_new(). The resulting object must be cleaned up via UA_NodeClass_delete(out::Ptr{UA_NodeClass}) after its use.

Open62541.UA_Server_readNodeId — Function

UA_Server_readNodeId(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_NodeId})

Uses the Server API to read the value of the attribute NodeId from the NodeId nodeId located on server server. The result is saved into out.

Note that memory for out must be allocated by C before using this function. This can be accomplished with out = UA_NodeId_new(). The resulting object must be cleaned up via UA_NodeId_delete(out::Ptr{UA_NodeId}) after its use.

Open62541.UA_Server_readSymmetric — Function

UA_Server_readSymmetric(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_Boolean})

Uses the Server API to read the value of the attribute Symmetric from the NodeId nodeId located on server server. The result is saved into out.

Note that memory for out must be allocated by C before using this function. This can be accomplished with out = UA_Boolean_new(). The resulting object must be cleaned up via UA_Boolean_delete(out::Ptr{UA_Boolean}) after its use.

Open62541.UA_Server_readValue — Function

UA_Server_readValue(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_Variant})

Uses the Server API to read the value of the attribute Value from the NodeId nodeId located on server server. The result is saved into out.

Note that memory for out must be allocated by C before using this function. This can be accomplished with out = UA_Variant_new(). The resulting object must be cleaned up via UA_Variant_delete(out::Ptr{UA_Variant}) after its use.

Open62541.UA_Server_readValueRank — Function

UA_Server_readValueRank(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_Int32})

Uses the Server API to read the value of the attribute ValueRank from the NodeId nodeId located on server server. The result is saved into out.

Note that memory for out must be allocated by C before using this function. This can be accomplished with out = UA_Int32_new(). The resulting object must be cleaned up via UA_Int32_delete(out::Ptr{UA_Int32}) after its use.

Open62541.UA_Server_readWriteMask — Function

UA_Server_readWriteMask(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, out::Ptr{UA_UInt32})

Uses the Server API to read the value of the attribute WriteMask from the NodeId nodeId located on server server. The result is saved into out.

Note that memory for out must be allocated by C before using this function. This can be accomplished with out = UA_UInt32_new(). The resulting object must be cleaned up via UA_UInt32_delete(out::Ptr{UA_UInt32}) after its use.

Writing to nodes:

Open62541.UA_Server_writeAccessLevel — Function

UA_Server_writeAccessLevel(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_Byte})

Uses the Server API to write the value new_val to the attribute AccessLevel of the NodeId nodeId located on the server.

Open62541.UA_Server_writeArrayDimensions — Function

UA_Server_writeArrayDimensions(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_Variant})

Uses the Server API to write the value new_val to the attribute ArrayDimensions of the NodeId nodeId located on the server.

Open62541.UA_Server_writeBrowseName — Function

UA_Server_writeBrowseName(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_QualifiedName})

Uses the Server API to write the value new_val to the attribute BrowseName of the NodeId nodeId located on the server.

Open62541.UA_Server_writeDataType — Function

UA_Server_writeDataType(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_NodeId})

Uses the Server API to write the value new_val to the attribute DataType of the NodeId nodeId located on the server.

Open62541.UA_Server_writeDataValue — Function

UA_Server_writeDataValue(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_DataValue})

Uses the Server API to write the value new_val to the attribute DataValue of the NodeId nodeId located on the server.

Open62541.UA_Server_writeDescription — Function

UA_Server_writeDescription(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_LocalizedText})

Uses the Server API to write the value new_val to the attribute Description of the NodeId nodeId located on the server.

Open62541.UA_Server_writeDisplayName — Function

UA_Server_writeDisplayName(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_LocalizedText})

Uses the Server API to write the value new_val to the attribute DisplayName of the NodeId nodeId located on the server.

Open62541.UA_Server_writeEventNotifier — Function

UA_Server_writeEventNotifier(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_Byte})

Uses the Server API to write the value new_val to the attribute EventNotifier of the NodeId nodeId located on the server.

Open62541.UA_Server_writeExecutable — Function

UA_Server_writeExecutable(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_Boolean})

Uses the Server API to write the value new_val to the attribute Executable of the NodeId nodeId located on the server.

Open62541.UA_Server_writeHistorizing — Function

UA_Server_writeHistorizing(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_Boolean})

Uses the Server API to write the value new_val to the attribute Historizing of the NodeId nodeId located on the server.

Open62541.UA_Server_writeInverseName — Function

UA_Server_writeInverseName(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_LocalizedText})

Uses the Server API to write the value new_val to the attribute InverseName of the NodeId nodeId located on the server.

Open62541.UA_Server_writeIsAbstract — Function

UA_Server_writeIsAbstract(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_Boolean})

Uses the Server API to write the value new_val to the attribute IsAbstract of the NodeId nodeId located on the server.

Open62541.UA_Server_writeMinimumSamplingInterval — Function

UA_Server_writeMinimumSamplingInterval(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_Double})

Uses the Server API to write the value new_val to the attribute MinimumSamplingInterval of the NodeId nodeId located on the server.

Open62541.UA_Server_writeValue — Function

UA_Server_writeValue(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_Variant})

Uses the Server API to write the value new_val to the attribute Value of the NodeId nodeId located on the server.

Open62541.UA_Server_writeValueRank — Function

UA_Server_writeValueRank(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_Int32})

Uses the Server API to write the value new_val to the attribute ValueRank of the NodeId nodeId located on the server.

Open62541.UA_Server_writeWriteMask — Function

UA_Server_writeWriteMask(server::Ptr{UA_Server}, nodeId::Ptr{UA_NodeId}, new_val::Ptr{UA_UInt32})

Uses the Server API to write the value new_val to the attribute WriteMask of the NodeId nodeId located on the server.